[IMP] Important improvement in functionality of our extension of res_users class. The function copy(..) is not anymore active. Was deleted because of problems with SignUp in webshop

+++ equitania/eq_res_users_new_api.py
@@ -32,29 +32,29 @@ class eq_res_users_new_api(models.Model):
         We'll also migrate implementation of res_users from old api to new api 
     """
         
-        
-    @api.model
+    """
+    @api.model    
     def create(self, vals):
-        """
-            Create handler for our extenstion of res_users
-            - creates new user and check if there's existing res_partner with same email and sets correct link between user and partner
-            
-            @vals: values to be saved
-            @return: call of super function
-        """        
-        
-        # check if we can find at least one record in res_partner with same email as user provided
-        partner = self.env['res.partner'].search([('email', '=', vals['email']), ('customer', '=', True)])        
-                            
+                
         # default call - create new user and save automaticaly created new partner_id            
-        new_user =  super(eq_res_users_new_api, self).create(vals)            
-        wrong_partner_id = new_user.partner_id.id
-        
-        # yes, we have an existing partner, so there's no need to use new created one. just use existing partner and delete new one
-        if partner.id is not False:                                        
-            new_user.partner_id = partner.id                                                    # set existing partner
-                                    
-            wrong_partner = self.env['res.partner'].search([('id', '=', wrong_partner_id)])     # get new + automaticaly generated partner and delete him - we don't need him
-            wrong_partner.unlink() 
-             
+        new_user =  super(eq_res_users_new_api, self).create(vals)
+            
+        # check if we can find at least one record in res_partner with same email as user provided
+        if 'email' in vals:
+            print "-----------------------  yes, it's here"
+            print "------ vals['email']: ", vals['email']
+            
+            partner = self.env['res.partner'].search([('email', '=', vals['email']), ('customer', '=', True)])                                                        
+            wrong_partner_id = new_user.partner_id.id
+            
+            # yes, we have an existing partner, so there's no need to use new created one. just use existing partner and delete new one
+            if partner.id is not False:                                        
+                new_user.partner_id = partner.id                                                    # set existing partner
+                                        
+                wrong_partner = self.env['res.partner'].search([('id', '=', wrong_partner_id)])     # get new + automaticaly generated partner and delete him - we don't need him
+                wrong_partner.unlink() 
+        else:
+            print "------------------------- no --------"
+                     
         return new_user
+    """
